Add session kwarg to ModelicaSystem constructor
If the session kwarg is specified, the ModelicaSystem will use that
session instead of creating a new one. This is useful for creating a
ModelicaSystem for a model that is already loaded in an already existing
session.

Remove __del__ method from ModelicaSystem. Having the deleter there
meant that the deleter for the session would run multiple times if a
session had shared ownership. The deleter will still run when the
session's reference count reaches zero, or when it is garbage collected.