Use pathlib.Path in OMCSessionHelper
I'm pretty sure this also fixes a bug:
the only Exception I can see os.path.join() raise in _get_omc_path()
is when self.omhome is None, which shouldn't really happen (the
constructor checks it).

I think the intention was to check whether the file actually exists,
so I do that instead.