Fix modelica system (#271)
* [ModelicaSystem] simplify code in xmlparse()

* [ModelicaSystem] add ModelicaSystemError() for unhandled final else cases

there are a lot of functions which check the type of the input; this is
done by if ... elif ... - however, invalid input is not catched, i.e. there
is *NO* return value define (would be None) if the input is not matching
any of the if branches

* [ModelicaSystem.getSolution()] do not try to continue on error but fail

Rule: fail early, fail hard - tell the user that something is wrong!

In this case, the user ask for the solution but could get None (= plain
'return') - this would case hard to track errors later (if verbose==False
and raiseerrors==False)

* [ModelicaSystem] remove redundant parentheses (type hint by PyCharm)